+++ src/components/UI/button/Button.tsx
@@ -1,5 +1,5 @@
 import { Component } from 'react';
-import { ButtonProps } from 'src/components/button/types';
+import { ButtonProps } from 'src/components/UI/button/types';
 
 export default class Button extends Component<ButtonProps> {
   render(): JSX.Element {
